fix test imports
The changes made for viabilize testing had broken the code imports
because relative imports were not working running pytest and to solve
that I had added 'src.' before the not working imports. However, I
hadnt' realided this broke the actual running. So, I removed the 'src.'
from where I changed before and created a conftes.py file. In this new
file I add the src directory to the sys.path, which allows me to run
the tests with no problem.