Remove error when upload folder doesn't exist
If the upload folder didn't exist the program would exit thrwing an
error. However, I think it makes more sense just create the folder. So,
for now the folder will be automatically created but in the future
maybe I'll add a dialog asking the user to chose a path and create the
folder