Add ConfigName class
The usual way of accessing a value in a ConfigParser involves passing
strings that represent the section and the value to a get function. So,
I was using hard coded strings to access the configs whenever I needed.
However it comes at the cost of increasing the amount of work if I ever
come to need changing one of this names. Then, I've created the
ConfigName class which holds constants representing the config sections
and values