Add ConfigWrapper
Python's ConfigParser is a good and simple to be used but I think it's
anoying have to remember the section and config names and have to use
this two informations to access the value when most of the time (to me)
the config nams are unique. Also, mapping each config to a constant
seems to be doing the same job twice. That's where ConfigWrapper gets
in.

A ConfigWrapper instance will have accessor methods for each config. The
method names are based on the config names and in case of duplicated
names they will also carry the section name. Consider the example bellow

[UPLOAD]
default-dir=/tmp/dir1

[DOWNLOAD]
default-dir=/tmp/dir2
ASK=False

For each of this configs a getter and a setter will be created. So we'll
have:

config_wrapper.get_ask()
config_wrapper.set_ask()
config_wrapper.get_upload_default_dir()
config_wrapper.get_download_default_dir()

and so on.